tests/wpt/tests/html/canvas/element/filters/2d.filter.layers.colorMatrix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.filter.layers.colorMatrix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.filter.layers.colorMatrix</h1>
<p class="desc">Test the functionality of ColorMatrix filters</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test the functionality of ColorMatrix filters");
_addTest(function(canvas, ctx) {

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix', values: undefined}}); });

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix', values: 'foo'}}); });

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix', values: null}}); });

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix', values: [1, 2, 3]}}); });

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix',
     values: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18,
              19, 'a']}}); });

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'colorMatrix',
     values: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18,
              19, Infinity]}}); });

  ctx.fillStyle = '#f00';
  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'hueRotate', values: 0}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 255,0,0,255, 2);

  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'hueRotate', values: 90}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 0,91,0,255, 2);

  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'hueRotate', values: 180}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 0,109,109,255, 2);

  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'hueRotate', values: 270}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 109,18,255,255, 2);

  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'saturate', values: 0.5}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 155,27,27,255, 2);

  ctx.clearRect(0, 0, 100, 50);
  ctx.beginLayer({filter:
    {name: 'colorMatrix', type: 'luminanceToAlpha'}});
  ctx.fillRect(0, 0, 100, 50);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 0,0,0,54, 2);

  ctx.beginLayer({filter:
    {name: 'colorMatrix', values: [
      0, 0, 0, 0, 0,
      1, 1, 1, 1, 0,
      0, 0, 0, 0, 0,
      0, 0, 0, 1, 0
    ]}});
  ctx.fillRect(0, 0, 50, 25);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(50, 0, 50, 25);
  ctx.fillStyle = '#00f';
  ctx.fillRect(0, 25, 50, 25);
  ctx.fillStyle = '#fff';
  ctx.fillRect(50, 25, 50, 25);
  ctx.endLayer();
  _assertPixelApprox(canvas, 10,10, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 60,10, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 10,30, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 60,30, 0,255,0,255, 2);

});
</script>
